+++ backend/api/management/commands/generate_fixtures.py
@@ -70,7 +70,7 @@ class Command(BaseCommand):
             category=ComputeTask.Category.TASK_TRAIN,
             status=ComputeTask.Status.STATUS_DONE,
         )
-        model_1 = factory.create_model(train_task_1, category=Model.Category.MODEL_SIMPLE)
+        model_1 = factory.create_model(train_task_1, category=Model.Category.MODEL_SIMPLE, identifier="model")
         train_task_2 = factory.create_computetask(
             doing_cp,
             simple_algo,
@@ -79,7 +79,7 @@ class Command(BaseCommand):
             category=ComputeTask.Category.TASK_TRAIN,
             status=ComputeTask.Status.STATUS_DONE,
         )
-        model_2 = factory.create_model(train_task_2, category=Model.Category.MODEL_SIMPLE)
+        model_2 = factory.create_model(train_task_2, category=Model.Category.MODEL_SIMPLE, identifier="model")
         factory.create_computetask(
             doing_cp,
             aggregate_algo,
@@ -103,7 +103,7 @@ class Command(BaseCommand):
             category=ComputeTask.Category.TASK_TRAIN,
             status=ComputeTask.Status.STATUS_DONE,
         )
-        model_3 = factory.create_model(train_task, category=Model.Category.MODEL_SIMPLE)
+        model_3 = factory.create_model(train_task, category=Model.Category.MODEL_SIMPLE, identifier="model")
         test_task = factory.create_computetask(
             done_cp,
             metric_algo,
@@ -128,7 +128,7 @@ class Command(BaseCommand):
             category=ComputeTask.Category.TASK_COMPOSITE,
             status=ComputeTask.Status.STATUS_DONE,
         )
-        model_4 = factory.create_model(composite_task, category=Model.Category.MODEL_HEAD)
+        model_4 = factory.create_model(composite_task, category=Model.Category.MODEL_HEAD, identifier="local")
         failed_task = factory.create_computetask(
             failed_cp,
             metric_algo,